+++ client/src/redux/slices/hierarchySlice.ts
@@ -1,9 +1,13 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
+import IEntity from '../../types/data/IEntity'
+
 export interface IHierarchy {
   currentPageLength: number
   previousPageLength: number
   defaultDisplayLength: number
+  origin: IEntity | null
+  fullscreen: boolean
 }
 
 const DEFAULT_LENGTH = 5
@@ -11,6 +15,8 @@ const initialState: IHierarchy = {
   currentPageLength: DEFAULT_LENGTH,
   previousPageLength: DEFAULT_LENGTH,
   defaultDisplayLength: DEFAULT_LENGTH,
+  origin: null,
+  fullscreen: false,
 }
 
 export const hierarchySlice = createSlice({
@@ -41,10 +47,22 @@ export const hierarchySlice = createSlice({
         state.defaultDisplayLength,
       )
     },
+    addOrigin: (state, action: PayloadAction<{ value: IEntity }>) => {
+      const { value } = action.payload
+      state.origin = value
+    },
+    addFullscreen: (
+      state,
+      action: PayloadAction<{ isFullscreen: boolean }>,
+    ) => {
+      const { isFullscreen } = action.payload
+      state.fullscreen = isFullscreen
+    },
     reset: () => initialState,
   },
 })
 
-export const { addShowMore, addShowLess, reset } = hierarchySlice.actions
+export const { addShowMore, addShowLess, addOrigin, addFullscreen, reset } =
+  hierarchySlice.actions
 
 export default hierarchySlice.reducer